sim submodule - move sim.py to own submodule - move write_and_run to sim.py - modify imports
