+++ nlmod/__init__.py
@@ -6,7 +6,7 @@
 
 import os
 
-from . import mdims, gwf, modpath, read, util, visualise
+from . import gwf, mdims, modpath, read, sim, util, visualise
 from .mdims import *
 from .version import __version__
 from .visualise import plots as plot
