+++ nlmod/gwf/__init__.py
@@ -1,4 +1,3 @@
-from .sim import *
 from .gwf import *
 from .constant_head import *
 from .horizontal_flow_barrier import *
